requirements.txt
Django~=2.2.4
djangorestframework~=3.10.2
djangorestframework_simplejwt==4.4.0
PyJWT~=1.7.1
djoser==2.0.3